feat: Don't have format depend on the main crate
