+++ format/Cargo.toml
@@ -17,7 +17,6 @@ itertools = "0.7"
 codespan = "0.2"
 
 gluon_base = { path = "../base", version = "0.9.4" } # GLUON
-gluon = { path = "..", version = "0.9.4" } # GLUON
 
 [dev-dependencies]
 env_logger = "0.6"
